Clarify conflating terms (post vs post)
My students had a few difficulties that came from the fact that they
didn't understand how the `POST` on the `request` object was related
to a blog `Post`. This was made clear once I explained the different
types of HTTP request (GET, POST etc.) and that the conflating terms
were just a coincidence.