Explain where missing posts go in deployment
Chat messages reveal that developers are shocked that data is not copied from their test system to deployment.  This patch anticipates the question and explains a little bit of what is going on behind the scenes with MySQL.  Also, the general flow of develop-test-commit-deploy was never explained and it seemed appropriate to quickly summarize it before detailing out the initial setup.