Don't set published_date when creating new post.
Actually, for this code to work well with the first Extentions homework (http://djangogirls.gitbooks.io/django-girls-tutorial-extensions/content/homework/index.html) it's probably best to not set the published_date here. Otherwise it won't be possible to create drafts, except from the admin site.

Sorry for all the back and forth turns with this publish_date thing.