Simplified the future workflow section
Heavy rewrites of the Future Workflow section to use consistent terms, get right to the point, and improve clarity.  Also added a section on the app-application-program-code nouns that keep getting pushed around.  I was afraid that the interchangeable use had never been directly addressed and the deploy section is where that becomes obvious, given that we check in code for our application, but Python code is not the only element in "the code."