Removed 8 part checklist, tightened up .gitignore
As requested by @olasitarska, the 8-step scary checklist is now a simple paragraph that connects the local build, GitHub, and PythonAnywhere.  Meanwhile, being a little more clear about the db.sqlite3 file being ignored on purpose and what that means later on.  There is still some ambiguity as to whether or not the deployed application actually uses MySQL on PythonAnywhere, so the text doesn't commit the rest of the tutorial to one technique or the other.