ALLOWED_HOSTS = ['*']
Starting in Django 1.8.16, the ALLOWED_HOSTS setting is checked even if `DEBUG` is `True`, so folks going through this tutorial will need to modify it.  I debated whether or not to set it to `['*']` in the `settings.py` section, or to go back and edit it during the deployment section.  This seemed like the easiest thing to do in the interest of getting started.