Add link to command window page
At the end of this page, it says to open a second command window. This can be tricky on Windows, as some versions switch you to existing windows instead of opening new windows. So, add a link to the Intro to Command Line page, where there are instructions for how to get around this.