Add more resources to "What's next?"
After coaching a number of sessions over the last couple of years
it's pretty clear that people come to the Django Girls tutorial
for all kinds of reasons – maybe they want to create a website for
their small business, maybe they want to use Python to help crunch
data for their PhD – the audience is _really_ broad. With that in
mind I wanted to flesh out and restructure the "What's next?"
section of the tutorial to better help people find resources that
are relevant to their interests so that they can better build on
their experience of the tutorial.

The only resource I've removed from the original page is the book
_Two Scoops of Django_ – I am a huge admirer of this book but I
believe it is far too advanced to give to someone coming straight
out of the Django Girls tutorial, and it costs money as well,
unlike the other resources on the page.

I'm on the fence about adding links to Coursera and edX since
there are paid-for components to their offerings, but I think
they could still provide value to our audience.