saner punctuation for parenthesized full sentences
When a full (and without the parenthesized part already complete)
sentence is followed by a parenthesized full complete sentence, then end
the former with a period before the parenthesis and end the latter with
a period within the parenthesis (before the closing brace). Also, start
the parenthesized sentence with a capital letter.

Same for stand-alone parenthesized sentences.