Changed h1 to h2
Chapter: Extending your application

This line

We want to have a link from a post's title in the post list to the post's detail page. Let's change <h1><a href="">{{ post.title }}</a></h1> so that it links to the post's detail page:

Should be changed to 

We want to have a link from a post's title in the post list to the post's detail page. Let's change <h2><a href="">{{ post.title }}</a></h2> so that it links to the post's detail page: