Clarify calling functions
The original explanation says that we re-write the name of the example
function at the bottom of the file "because Python reads the file from
top to bottom, so we need to re-write the name of the function." I find
that explanation a bit confusing myself.

I think two things were conflated here:
1) The difference between defining and running a function
2) Explaining that Python reads files from top to bottom, and needs
functions to be defined before they can be called

I tried to more explicitly explain that what we're doing is calling a
function we just defined, and break these concepts apart more.