Update instructions.md
Added an easier way to bring up the system information app on a Win10 machine, as there is no Pause/Break key on Microsoft Laptops. As a side note - what I am seeing on the published website does not match the Master Branch. Perhaps someone could deploy it before Sat?