Add example of method result comparison
As the following paragraph states "As you can see, not only can Python compare numbers, but it can also compare method results." I think it makes sense to include a comparison of method results in the example itself. Since len() has been used earlier in the chapter its behavior should be familiar to the reader already.