Fix: Command-line Windows tooltip incorrectly built
Because of missing newlines in some languages, the `Opening: Windows`
tooltip is incorrectly created as an empty section (included in previous
`h1` tag).

This change follows the same format the EN language has, where the
`include` is done in a separate line, and each of the included files has
a newline at the beginning, to avoid this issue in the future.