Ignore the default branch name problem
With older versions of git the command to rename the branch fails if
there is no branch yet. We're also making the tutorial harder than it
needs to be.

This ignores it and simply uses the default by pushing HEAD instead of
main. GitHub is then smart enough to use that as default. Future clones
will also simply copy the default branch from GitHub.