remove suggestions to use Jupyter Lab workspace
as well as all mentions of how to open
a command line window or Python prompt in it,
as (for this tutorial) there's probably no advantage
to use Jupyter Lab in RunCode
over using RunCode's default VSCode based workspaces,
but indeed some disadvantages, as #1805 mentions.
(Fewer unneeded options to choose from also simplifies
the work of coaches and of those maintaining the tutorial.)

Also, completely remove the special case for RunCode users
in the "Python prompt" section, as at that point,
they've already been told to follow Linux instructions.

fixes #1805