Run GitHub Actions on Ubuntu 24.04
On Ubuntu the file /usr/bin/rename is managed through alternatives
because there are multiple implementations. It appears that
awalsh128/cache-apt-pkgs-action doesn't update alternatives when a
package is cached which means /usr/bin/rename isn't present.

The latest image will change to 24.04 on between December and January
but on 24.04 there is rename.ul from util-linux which also provides the
rename. That avoids the need to rely on awalsh128/cache-apt-pkgs-action
for it and the alternatives interaction.