Remove instructions to check if windows is 32-bit or 64-bit.
Windows 10 was the last version of windows to support 32-bit, that dates back to 10 years ago now, and the latest version, Windows 11, only has 64-bit, since 2021.  most manufacturers stopped had stopped building 32-bit computers by 2017 (https://www.quora.com/Are-32-bit-computers-still-being-produced-as-of-2017)

This change proposes we remove the big, fiddly chunk of instructions for "how to check whether you have 32-bit windows",

but!  it replaces them with a mention of using 32-bit in the already-existing "what to do if you have any problems" paragraph, later on.

my hope is that this makes life easier for the 90% of users on 64-bit windows, whilst not making things that much worse for people with older PCs.

am aware that other cultural contexts may be v different in respect, where there is a lot of older hardware around!