Tell people to create a GitHub Personal Access Token up-front
When we get to the `git push` stage,
the tutorial currently tells people to attempt to use their normal
github username and password.

It goes on to say "oh you will probably get an error",
and tells people to create a PAT instead.

GitHub actually removed username & password auth in 2021.

So let's just tell people to use a PAT up-front.

ALSO: the tutorial used to tell people to follow the instructions on
GitHub, but then also included its own instructions, inline.
So I've removed them in favour of making github the single source of
truth.